myr-2376: restructure server api
